refactor: move oidc test mocks into top oidc test class
