+++ allauth/socialaccount/providers/openid_connect/tests.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import json
 from unittest import TestSuite
 
 from allauth.socialaccount.providers.openid_connect.provider import (
@@ -7,52 +6,13 @@ from allauth.socialaccount.providers.openid_connect.provider import (
     provider_classes,
 )
 from allauth.socialaccount.tests import OpenIDConnectTests
-from allauth.tests import Mock, MockedResponse, TestCase, patch
+from allauth.tests import TestCase
 
 from ... import app_settings
 
 
 class OpenIDConnectTestsBase(OpenIDConnectTests):
     provider_id = None
-    oidc_info_content = {
-        "authorization_endpoint": "/login",
-        "userinfo_endpoint": "/userinfo",
-        "token_endpoint": "/token",
-    }
-    userinfo_content = {
-        "picture": "https://secure.gravatar.com/avatar/123",
-        "email": "ness@some.oidc.server.onett.example",
-        "id": 1138,
-        "sub": 2187,
-        "identities": [],
-        "name": "Ness",
-    }
-    extra_data = {
-        "picture": "https://secure.gravatar.com/avatar/123",
-        "email": "ness@some.oidc.server.onett.example",
-        "id": 2187,
-        "identities": [],
-        "name": "Ness",
-    }
-
-    def setUp(self):
-        super(OpenIDConnectTestsBase, self).setUp()
-        patcher = patch(
-            "allauth.socialaccount.providers.openid_connect.views.requests",
-            get=Mock(side_effect=self._mocked_responses),
-        )
-        self.mock_requests = patcher.start()
-        self.addCleanup(patcher.stop)
-
-    def get_mocked_response(self):
-        # Enable test_login in OAuth2TestsMixin, but this response mock is unused
-        return True
-
-    def _mocked_responses(self, url, *args, **kwargs):
-        if url.endswith("/.well-known/openid-configuration"):
-            return MockedResponse(200, json.dumps(self.oidc_info_content))
-        elif url.endswith("/userinfo"):
-            return MockedResponse(200, json.dumps(self.userinfo_content))
 
     def test_oidc_base_and_provider_settings_sync(self):
         # Retrieve settings via this OpenID Connect server's specific provider ID
